Rename some files to match their content
