+++ src/SignalHandler.cpp
@@ -1,0 +1,75 @@
+// Copyright (c) 2005 - 2015 Settlers Freaks (sf-team at siedler25.org)
+//
+// This file is part of Return To The Roots.
+//
+// Return To The Roots is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 2 of the License, or
+// (at your option) any later version.
+//
+// Return To The Roots is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Return To The Roots. If not, see <http://www.gnu.org/licenses/>.
+
+#include "defines.h" // IWYU pragma: keep
+#ifdef _WIN32
+#   include <windows.h>
+#else
+#   include <cstdlib>
+#endif // _WIN32
+#include "SignalHandler.h"
+
+#include "GlobalVars.h"
+#include <iostream>
+
+/**
+ *  Signal-Handler
+ */
+#ifdef _WIN32
+BOOL WINAPI HandlerRoutine(DWORD dwCtrlType)
+{
+    switch(dwCtrlType)
+    {
+        case CTRL_BREAK_EVENT:
+        case CTRL_CLOSE_EVENT:
+        case CTRL_C_EVENT:
+        {
+            GLOBALVARS.notdone = false;
+            return TRUE;
+        } break;
+    }
+    return FALSE;
+}
+#else
+bool killme = false;
+void HandlerRoutine(int sig)
+{
+
+    int c;
+    if(sig != SIGINT)
+        return;
+    else
+    {
+        if(!killme)
+            std::cout << "Wollen Sie das Programm beenden (j/n) : ";
+        else
+            std::cout << "Wollen Sie das Programm killen (j/n) : ";
+
+        c = getchar();
+        if(c == 'j' || c == 'y' || c == 1079565930)
+        {
+            if(killme)
+                exit(1);
+
+            killme = true;
+            GLOBALVARS.notdone = false;
+        }
+        else
+            return;
+    }
+}
+#endif // _WIN32
