+++ src/gameTypes/MapInfo.h
@@ -18,8 +18,8 @@
 #ifndef MapInfo_h__
 #define MapInfo_h__
 
-#include "GameProtocol.h"
 #include "gameTypes/CompressedData.h"
+#include "gameTypes/MapType.h"
 #include "helpers/Deleter.h"
 #include <boost/interprocess/smart_ptr/unique_ptr.hpp>
 #include <string>
