+++ src/gameTypes/MapType.h
@@ -1,4 +1,4 @@
-// Copyright (c) 2005 - 2015 Settlers Freaks (sf-team at siedler25.org)
+// Copyright (c) 2005 - 2016 Settlers Freaks (sf-team at siedler25.org)
 //
 // This file is part of Return To The Roots.
 //
@@ -15,27 +15,13 @@
 // You should have received a copy of the GNU General Public License
 // along with Return To The Roots. If not, see <http://www.gnu.org/licenses/>.
 
-#include "defines.h" // IWYU pragma: keep
-#include "MapInfo.h"
-#include "Savegame.h"
+#ifndef MapType_h__
+#define MapType_h__
 
-MapInfo::MapInfo()
+enum MapType
 {
-    Clear();
-}
+    MAPTYPE_OLDMAP = 0,
+    MAPTYPE_SAVEGAME
+};
 
-MapInfo::~MapInfo()
-{}
-
-void MapInfo::Clear()
-{
-    type = MAPTYPE_OLDMAP;
-    title.clear();
-    filepath.clear();
-    luaFilepath.clear();
-    mapData.Clear();
-    luaData.Clear();
-    mapChecksum = 0;
-    luaChecksum = 0;
-    savegame.reset();
-}
+#endif // MapType_h__
